requirements.txt
# --- Web Server & API ---
fastapi
uvicorn
requests
python-dotenv
qstash

# --- Scheduling & Logic ---
apscheduler
feedparser==6.0.11
numpy

# --- AI Content (Script & Voice) ---
openai
edge-tts

# --- Image Processing & Upscaling ---
Pillow

# --- Video Rendering ---
moviepy==2.1.1
ffmpeg-python

# --- Cloud Storage & Fallbacks ---
cloudinary
huggingface_hub